moved a whole bunch of stuff to GameConstants.cs, yay! no more public anything unless it's prefabs/objects!
